LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/compile-event-handler-lexical-scopes-form-owner-expected.txt


div
x-foo

PASS <input> has a form owner
PASS <img> has a form owner
PASS <div> doesn't have a form owner
FAIL form-associated <x-foo> has a form owner assert_equals: expected (object) object "[object HTMLFormControlsCollection]" but got (string) "global_elements"
